analysis-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
numpy>=1.26.0,<2.0.0
pandas>=2.2.0
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
chromadb==0.4.22
ta==0.11.0
python-dotenv==1.0.0
openai==1.50.0
requests==2.32.0
